zfpman API completed, data not verified yet

+++ examples/writeZfp.cc
@@ -18,15 +18,15 @@ int main(){
 
     std::unique_ptr<DataMan> man = func();
 
-    vector<uint64_t> putshape;
+    vector<size_t> putshape;
     putshape.assign(3,0);
 
-    vector<uint64_t> varshape;
+    vector<size_t> varshape;
     varshape.push_back(0);
     varshape.push_back(0);
     varshape.push_back(0);
 
-    vector<uint64_t> offset;
+    vector<size_t> offset;
     offset.push_back(0);
     offset.push_back(0);
     offset.push_back(0);
